更换Compare模块的位置
